merge main a mi branch

+++ backend/api/views.py
@@ -1,3 +1,4 @@
+from rest_framework.exceptions import ValidationError
 from .serializers import (
     ProyectoSerializer,
     RolAsignadoSerializer,
@@ -20,59 +21,157 @@ def proyectos(request, proyect_id=None):
     if request.method == "POST":
         # Crea el proyecto
         data = JSONParser().parse(request)
-        serializer = ProyectoSerializer(data=data)
-        if serializer.is_valid():
-            serializer.save()
-            return JsonResponse(serializer.data, status=201)
-        return JsonResponse(serializer.errors, status=400, safe=False)
+        try:
+            proy_seri = ProyectoSerializer(data=data)
+            proy_seri.is_valid(raise_exception=True)
+            proy = proy_seri.save()
+
+            # Crea el rol scrum master en el proyecto
+            rol_seri = RolSerializer(
+                data={
+                    "nombre": "Scrum Master",
+                    "proyecto": proy.id,
+                    "permisos": [
+                        1,
+                        2,
+                        3,
+                        4,
+                        5,
+                        6,
+                        7,
+                        8,
+                        9,
+                        10,
+                        11,
+                        12,
+                        13,
+                        14,
+                        15,
+                        16,
+                        17,
+                        18,
+                        19,
+                        20,
+                        21,
+                    ],
+                }
+            )
+
+            rol_seri.is_valid(raise_exception=True)
+            scrum_rol = rol_seri.save(id=proy.id)
+
+            # Asigna el rol scrum mastes al miembro en la posicion 0
+            rol_asignado_seri = RolAsignadoSerializer(
+                data={
+                    "rol": scrum_rol.id,
+                    "usuario": data["scrumMasterId"],
+                    "proyecto": proy.id,
+                }
+            )
+            rol_asignado_seri.is_valid(raise_exception=True)
+            rol_asignado_seri.save(id=proy.id)
+
+            # Crea el rol Developer en el proyecto
+            rol_seri = RolSerializer(
+                data={
+                    "nombre": "Developer",
+                    "proyecto": proy.id,
+                    "permisos": [
+                        1,
+                        6,
+                        8,
+                        10,
+                    ],
+                }
+            )
+
+            rol_seri.is_valid(raise_exception=True)
+            scrum_rol = rol_seri.save()
+
+            # Crea el rol Product Owner en el proyecto
+            rol_seri = RolSerializer(
+                data={
+                    "nombre": "Product Owner",
+                    "proyecto": proy.id,
+                    "permisos": [
+                        1,
+                        2,
+                        5,
+                        6,
+                        7,
+                        9,
+                    ],
+                }
+            )
+
+            rol_seri.is_valid(raise_exception=True)
+            scrum_rol = rol_seri.save()
+
+            return JsonResponse(proy_seri.data, status=201, safe=False)
+        except ValidationError as e:
+            return JsonResponse(e.detail, status=400, safe=False)
+        except Exception as e:
+            return JsonResponse(e, safe=False)
 
     elif request.method == "GET":
         if proyect_id != None:
+            # Trae un proyecto por su proyect_id
             try:
                 p = Proyecto.objects.get(id=proyect_id)
-                serializer = ProyectoSerializer(p)
-                return JsonResponse(serializer.data, safe=False)
+                proy_seri = ProyectoSerializer(p)
+                proy_data = proy_seri.data
+                scrum_obj = RolAsignado.objects.get(
+                    id=proyect_id
+                )  ## obtiene el RolAsignado con el id del proyecto, el del scrum master
+                scrum_data = RolAsignadoSerializer(scrum_obj).data
+                proy_data.update({"scrumMaster": scrum_data})
+                return JsonResponse(proy_data, safe=False)
             except Usuario.DoesNotExist:
                 return HttpResponseNotFound()
         else:
+            # Trae todos los proyectos
             p = Proyecto.objects.all()
-            serializer = ProyectoSerializer(p, many=True)
-            return JsonResponse(serializer.data, safe=False)
+            proy_seri = ProyectoSerializer(p, many=True)
+            return JsonResponse(proy_seri.data, safe=False)
 
     elif request.method == "DELETE":
+        # Elimina un proyecto
         p = Proyecto.objects.get(id=proyect_id)
         p.delete()
         return JsonResponse(True, safe=False, status=204)
 
-
     elif request.method == "PUT":
         if proyect_id:
+            # Editar la informacion de un proyecto
             data = JSONParser().parse(request)
 
             rol = Proyecto.objects.get(id=proyect_id)
 
-            serializer = ProyectoSerializer(rol, data=data, partial=True)
+            proy_seri = ProyectoSerializer(rol, data=data, partial=True)
 
-            if serializer.is_valid():
-                # Obtiene el id del Rol para vincular
-                serializer.save()
-                return JsonResponse(serializer.data, status=200)
-            return JsonResponse(serializer.errors, status=400, safe=False)
+            if proy_seri.is_valid():
+                proy_seri.save()
+                return JsonResponse(proy_seri.data, status=200)
+            return JsonResponse(proy_seri.errors, status=400, safe=False)
         return HttpResponseBadRequest("Falta proyect_id")
 
 
 def usuarios(request, user_id=None):
     if request.method == "POST":
-
+        # Crea un Nuevo usuario
         data = JSONParser().parse(request)
         serializer = UsuarioSerializer(data=data)
         if serializer.is_valid():
-            serializer.save()
+            if data.get("id"):
+                serializer.save(id=data.get("id"))
+            else:
+                serializer.save()
             return JsonResponse(serializer.data, status=201)
         return JsonResponse(serializer.errors, status=400, safe=False)
 
     elif request.method == "GET":
         if not user_id and not request.GET.get("email"):
+            # Trae Todos los usuarios de la base de datos
             u = Usuario.objects.all()
             serializer = UsuarioSerializer(u, many=True)
             return JsonResponse(serializer.data, safe=False)
@@ -80,18 +179,20 @@ def usuarios(request, user_id=None):
         if not request.GET.get("email") and not user_id:
             return HttpResponseBadRequest("Falta el mail o el user en el body")
         try:
+            # Trae Un usuario por su respectivo user_id o email
             if request.GET.get("email"):
                 u = Usuario.objects.get(email=request.GET.get("email"))
                 serializer = UsuarioSerializer(u)
             else:
                 u = Usuario.objects.get(id=user_id)
                 serializer = UsuarioSerializer(u)
-                print(serializer.data)
+
             return JsonResponse(serializer.data, safe=False)
         except Usuario.DoesNotExist:
             return HttpResponseNotFound()
     elif request.method == "PUT":
         if user_id:
+            # Edita la informacion de un Usuario
             data = JSONParser().parse(request)
 
             rol = Usuario.objects.get(id=user_id)
@@ -99,22 +200,21 @@ def usuarios(request, user_id=None):
             serializer = UsuarioSerializer(rol, data=data, partial=True)
 
             if serializer.is_valid():
-                # Obtiene el id del Rol para vincular
                 serializer.save()
                 return JsonResponse(serializer.data, status=200)
             return JsonResponse(serializer.errors, status=400, safe=False)
         return HttpResponseBadRequest("Falta user_id")
     elif request.method == "DELETE":
+        # Elimina un usuario
         u = Usuario.objects.get(id=user_id)
         u.delete()
         return JsonResponse(True, safe=False, status=204)
 
 
 def usuarios_proyectos(request, user_id):
-    # agregar y eliminar\
+    # trae todos los usuarios del proyecto
 
     if request.method == "GET":
-        # trae los proyectos del usuario
 
         try:
             p = Proyecto.objects.filter(miembros__id__contains=user_id)
@@ -125,8 +225,8 @@ def usuarios_proyectos(request, user_id):
 
 
 def proyectos_miembros(request, proyect_id, user_id=None):
-    # agregar y eliminar\
-    # proyect_id = request.GET.get("proyect_id")
+    # agregar y eliminar usuarios
+
     if request.method == "POST":
         if not user_id:
             return JsonResponse(False, status=400, safe=False)
@@ -141,6 +241,20 @@ def proyectos_miembros(request, proyect_id, user_id=None):
             p.miembros.add(u)
             p.save()
 
+            # Asgina el rol de Developer al usuario
+            pr = Rol.objects.filter(proyecto=proyect_id)
+            pr_seri = RolSerializer(pr, many=True)
+            for rol in pr_seri.data:
+                if rol["nombre"] == "Developer":
+                    RolAsignadoSerializer(
+                        data={
+                            "usuario": user_id,
+                            "proyecto": proyect_id,
+                            "rol": rol["id"],
+                        }
+                    )
+                    break
+
             return JsonResponse(True, status=201, safe=False)
         except Exception as e:
             return JsonResponse(str(e), status=400, safe=False)
@@ -153,13 +267,23 @@ def proyectos_miembros(request, proyect_id, user_id=None):
             u_list = []
             for id in serializer.data["miembros"]:
                 u = Usuario.objects.get(id=id)
-                u_list.append(UsuarioSerializer(u).data)
+                u_data = UsuarioSerializer(u).data
+                r = RolAsignado.objects.filter(proyecto=proyect_id, usuario=id)
+                rol_data = RolAsignadoSerializer(r, many=True).data
+                rol_data = rol_data[0] if len(rol_data) > 0 else None
+                u_data.update({"rol": rol_data})
+                u_list.append(u_data)
             return JsonResponse(u_list, safe=False)
 
         try:
             u = Usuario.objects.get(id=user_id)
+            user_data = UsuarioSerializer(u).data
+            r = RolAsignado.objects.filter(proyecto=proyect_id, usuario=user_id)
+            rol_data = RolAsignadoSerializer(r, many=True).data
+            rol_data = rol_data[0] if len(rol_data) > 0 else None
 
-            return JsonResponse(UsuarioSerializer(u).data, safe=False)
+            user_data.update({"rol": rol_data})
+            return JsonResponse(user_data, safe=False)
         except Usuario.DoesNotExist:
             return HttpResponseNotFound()
 
@@ -172,6 +296,15 @@ def proyectos_miembros(request, proyect_id, user_id=None):
             p = Proyecto.objects.get(id=proyect_id)
             p.miembros.remove(user_id)
             p.save()
+
+            # Elimina Roles asignados al proyecto
+            rolesAsignados = RolAsignado.objects.filter(
+                proyecto=proyect_id, usuario=user_id
+            )
+
+            for ra in rolesAsignados:
+                ra.delete()
+
             serializer = ProyectoSerializer(p)
             return JsonResponse(serializer.data, safe=False, status=204)
         except Usuario.DoesNotExist:
@@ -181,8 +314,8 @@ def proyectos_miembros(request, proyect_id, user_id=None):
 # api para manejar los roles
 def roles(request, proyect_id, rol_id=None):
     if request.method == "POST":
-        data = JSONParser().parse(request)
         # Crea un nuevo rol en el proyecto
+        data = JSONParser().parse(request)
         serializer = RolSerializer(
             data={
                 "nombre": data["nombre"],
@@ -191,12 +324,16 @@ def roles(request, proyect_id, rol_id=None):
             }
         )
         if serializer.is_valid():
-            # Obtiene el id del Rol para vincular
             serializer.save()
             return JsonResponse(serializer.data, status=201)
         return JsonResponse(serializer.errors, status=400, safe=False)
     elif request.method == "PUT":
         if rol_id:
+            # Editar un Rol
+            if rol_id == proyect_id:
+                return JsonResponse(
+                    "No se puede editar el Rol de Scrum Master", safe=False, status=400
+                )
             data = JSONParser().parse(request)
 
             rol = Rol.objects.get(id=rol_id)
@@ -217,6 +354,7 @@ def roles(request, proyect_id, rol_id=None):
             return JsonResponse(seri.data, safe=False)
 
         try:
+            # Trae un Rol por su id
             u = Rol.objects.get(id=rol_id)
             return JsonResponse(RolSerializer(u).data, safe=False)
         except Usuario.DoesNotExist:
@@ -226,6 +364,10 @@ def roles(request, proyect_id, rol_id=None):
         # En caso de DELETE elimina el rol  del proyecto
 
         if rol_id:
+            if rol_id == proyect_id:
+                return JsonResponse(
+                    "No se puede eliminar el Rol", safe=False, status=400
+                )
             try:
                 r = Rol.objects.get(proyecto=proyect_id, id=rol_id)
                 r.delete()
@@ -237,26 +379,48 @@ def roles(request, proyect_id, rol_id=None):
 
 
 # Api para asignar los roles a los usuarios
-def usuarios_proyectos_roles(request, proyect_id, user_id, rol_id=None):
+def proyectos_miembros_roles(request, proyect_id, user_id, rol_id=None):
 
     if request.method == "POST":
         # En caso de POST asigna un rol a un usuario
         if not rol_id:
             return HttpResponseBadRequest("Falta rol_id")
 
-        seri = RolAsignadoSerializer(
-            data={"usuario": user_id, "rol": rol_id, "proyecto": proyect_id}
-        )
-        if seri.is_valid():
+        try:
+            rolAsig = RolAsignado.objects.filter(proyecto=proyect_id, usuario=user_id)
+            for ra in rolAsig:
+                ra.delete()
+
+            if proyect_id == rol_id:
+                # Si es scrum master el rol asignado cambia
+                scrum = RolAsignado.objects.get(id=rol_id)
+                seri = RolAsignadoSerializer(
+                    scrum,
+                    data={
+                        "usuario": user_id,
+                    },
+                    partial=True,
+                )
+            else:
+                seri = RolAsignadoSerializer(
+                    data={"usuario": user_id, "rol": rol_id, "proyecto": proyect_id}
+                )
+            seri.is_valid(raise_exception=True)
             seri.save()
             return JsonResponse(seri.data, status=201)
-        return JsonResponse(seri.errors, status=400, safe=False)
+        except:
+            return JsonResponse(seri.errors, status=400, safe=False)
     elif request.method == "GET":
         # En caso de GET trae todos los roles del usuario en el proyecto
         try:
-            r = RolAsignado.objects.filter(proyecto=proyect_id, usuario=user_id)
-            seri = RolAsignadoSerializer(r, many=True)
-            return JsonResponse(seri.data, safe=False)
+            rolAsig = RolAsignado.objects.filter(proyecto=proyect_id, usuario=user_id)
+            rolAsi = rolAsig[0] if len(rolAsig) > 0 else None
+            seri = RolAsignadoSerializer(rolAsi)
+
+            rol = Rol.objects.get(id=seri.data["id"])
+            rol_seri = RolSerializer(rol)
+
+            return JsonResponse(rol_seri.data, safe=False)
         except Rol.DoesNotExist:
             return HttpResponseNotFound()
 
@@ -265,10 +429,10 @@ def usuarios_proyectos_roles(request, proyect_id, user_id, rol_id=None):
 
         if rol_id:
             try:
-                r = RolAsignado.objects.get(
+                rolAsig = RolAsignado.objects.get(
                     proyecto=proyect_id, usuario=user_id, rol=rol_id
                 )
-                r.delete()
+                rolAsig.delete()
 
                 return JsonResponse(True, safe=False, status=204)
             except RolAsignado.DoesNotExist:
@@ -283,7 +447,7 @@ def user_stories(request, proyect_id, us_id=None):
         return HttpResponseNotFound()
 
     if request.method == "POST":
-        # Crea el user storie
+        # Crea la user storie
         data = JSONParser().parse(request)
         data["proyecto"] = proyect_id
         serializer = USSerializer(data=data)
@@ -308,7 +472,7 @@ def user_stories(request, proyect_id, us_id=None):
     elif request.method == "DELETE":
         us = US.objects.get(id=us_id)
         us.delete()
-        return JsonResponse(True, status=204)
+        return JsonResponse(True, status=204, safe=False)
 
     elif request.method == "PUT":
         if us_id:
@@ -323,3 +487,67 @@ def user_stories(request, proyect_id, us_id=None):
                 return JsonResponse(serializer.data, status=200)
             return JsonResponse(serializer.errors, status=400, safe=False)
         return HttpResponseBadRequest("Falta us_id")
+
+
+def sprints(request, proyect_id, sprint_id=None):
+    try:
+        proyecto = Proyecto.objects.get(id=proyect_id)
+    except Proyecto.DoesNotExist:
+        return HttpResponseNotFound()
+
+    if request.method == "POST":
+        # Crea un sprint
+        data = JSONParser().parse(request)
+        data["proyecto"] = proyect_id
+        serializer = SprintSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return JsonResponse(serializer.data, status=201)
+        return JsonResponse(serializer.errors, status=400, safe=False)
+
+    elif request.method == "GET":
+        if sprint_id != None:
+            # Retorna un sprint
+            try:
+                spr = Sprint.objects.get(id=sprint_id)
+                serializer = SprintSerializer(spr)
+                return JsonResponse(serializer.data, safe=False)
+            except Sprint.DoesNotExist:
+                return HttpResponseNotFound()
+        else:
+            # Retorna los sprints de un proyecto
+            spr = Sprint.objects.filter(proyecto=proyecto)
+            serializer = SprintSerializer(spr, many=True)
+            return JsonResponse(serializer.data, safe=False)
+
+    elif request.method == "DELETE":
+        spr = Sprint.objects.get(id=sprint_id)
+        spr.delete()
+        return JsonResponse(True, status=204, safe=False)
+
+    elif request.method == "PUT":
+        if sprint_id:
+            data = JSONParser().parse(request)
+
+            spr = Sprint.objects.get(id=sprint_id)
+
+            serializer = SprintSerializer(spr, data=data, partial=True)
+
+            if serializer.is_valid():
+                serializer.save()
+                return JsonResponse(serializer.data, status=200)
+            return JsonResponse(serializer.errors, status=400, safe=False)
+        return HttpResponseBadRequest("Falta sprint_id")
+
+
+# Retorna los user stories de un sprint
+def sprints_user_stories(request, proyect_id, sprint_id):
+    try:
+        proyecto = Proyecto.objects.get(id=proyect_id)
+    except Proyecto.DoesNotExist:
+        return HttpResponseNotFound()
+    if request.method == "GET":
+        us = US.objects.filter(proyecto=proyecto, sprint=sprint_id)
+        serializer = USSerializer(us, many=True)
+        return JsonResponse(serializer.data, safe=False)
+    return HttpResponseBadRequest("Falta sprint_id")
